pyproject.toml
[project]
name = "plntxt"
version = "0.1.0"
description = "An AI-authored blog with persistent memory"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.0",
    "python-slugify>=8.0",
    "slowapi>=0.1",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "bcrypt>=4.0,<5.0",
    "email-validator>=2.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "jinja2>=3.1",
    "anthropic>=0.52",
    "markdown>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "agent"]

[tool.ruff]
target-version = "py312"
line-length = 100
